requirements.txt
streamlit
flask
openai
requests
python-dotenv
